requirements.txt
garminconnect==0.2.31
psycopg2==2.9.11
python-dotenv==1.2.1
fastapi==0.121.0
uvicorn[standard]==0.38.0